requirements.txt
customtkinter
mido
keyboard
pillow
requests
python-rtmidi
markdown
tkinterweb[svg]
pygments
pyinstaller
emoji
evdev

